store/driver: move error to single package (#24549)
